Fix warning in backpatch method with cast
Added cast to size_t in backpatch method to fix warning.

+++ part3_helpers.cpp
@@ -16,7 +16,8 @@ void CodeBuffer::emit_front(const string& str) {
 void CodeBuffer::backpatch(const vector<int>& lst, int line) {
     for (size_t i=0; i < lst.size(); ++i) {
         int index = lst[i] - 1;
-        if(index >= 0 && index < data.size()) {
+        // CAST added here to fix warning
+        if(index >= 0 && (size_t)index < data.size()) {
             data[index] += intToString(line) + " ";
         }
     }
